+++ agent/agent-plugins/apache-httpcomponents4/src/main/java/org/bithon/agent/plugin/apache/httpcomponents4/interceptor/DefaultClientConnectionOperator$OpenConnection.java
@@ -1,0 +1,63 @@
+/*
+ *    Copyright 2020 bithon.org
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.bithon.agent.plugin.apache.httpcomponents4.interceptor;
+
+import org.apache.http.HttpHost;
+import org.apache.http.conn.OperatedClientConnection;
+import org.apache.http.params.HttpParams;
+import org.apache.http.protocol.HttpContext;
+import org.bithon.agent.instrumentation.aop.context.AopContext;
+import org.bithon.agent.instrumentation.aop.interceptor.InterceptionDecision;
+import org.bithon.agent.instrumentation.aop.interceptor.declaration.AroundInterceptor;
+import org.bithon.agent.observability.tracing.context.ITraceSpan;
+import org.bithon.agent.observability.tracing.context.TraceContextFactory;
+import org.bithon.component.commons.tracing.Tags;
+
+import java.net.InetAddress;
+
+/**
+ * {@link org.apache.http.impl.conn.DefaultClientConnectionOperator#openConnection(OperatedClientConnection, HttpHost, InetAddress, HttpContext, HttpParams)}
+ *
+ * @author frank.chen021@outlook.com
+ * @date 2023/10/25 17:42
+ */
+public class DefaultClientConnectionOperator$OpenConnection extends AroundInterceptor {
+
+    @Override
+    public InterceptionDecision before(AopContext aopContext) {
+        ITraceSpan span = TraceContextFactory.newSpan("http-client");
+        if (span == null) {
+            return InterceptionDecision.SKIP_LEAVE;
+        }
+
+        HttpHost httpHost = aopContext.getArgAs(1);
+        aopContext.setSpan(span.method(aopContext.getTargetClass(), aopContext.getMethod())
+                               // Since this span does not propagate the tracing context to next hop,
+                               // it's not marked as SpanKind.CLIENT
+                               .tag(Tags.Http.CLIENT, "apache")
+                               .tag(Tags.Net.PEER, httpHost.getPort() == -1 ? httpHost.getHostName() : httpHost.getHostName() + ":" + httpHost.getPort())
+                               .start());
+
+        return InterceptionDecision.CONTINUE;
+    }
+
+    @Override
+    public void after(AopContext aopContext) {
+        ITraceSpan span = aopContext.getSpan();
+        span.tag(aopContext.getException()).finish();
+    }
+}
