+++ agent/agent-plugins/apache-httpcomponents5/src/main/java/org/bithon/agent/plugin/apache/httpcomponents5/interceptor/InternalHttpClient$DoExecute.java
@@ -1,0 +1,87 @@
+/*
+ *    Copyright 2020 bithon.org
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.bithon.agent.plugin.apache.httpcomponents5.interceptor;
+
+import org.apache.hc.client5.http.impl.classic.CloseableHttpResponse;
+import org.apache.hc.core5.http.ClassicHttpRequest;
+import org.apache.hc.core5.http.HttpHost;
+import org.apache.hc.core5.http.protocol.HttpContext;
+import org.bithon.agent.instrumentation.aop.context.AopContext;
+import org.bithon.agent.instrumentation.aop.interceptor.InterceptionDecision;
+import org.bithon.agent.instrumentation.aop.interceptor.declaration.AroundInterceptor;
+import org.bithon.agent.observability.metric.domain.http.HttpOutgoingMetrics;
+import org.bithon.agent.observability.metric.domain.http.HttpOutgoingMetricsRegistry;
+
+import java.util.HashSet;
+import java.util.Locale;
+import java.util.Set;
+
+/**
+ * Apache http component(client) interceptor for
+ * {@link org.apache.hc.client5.http.impl.classic.InternalHttpClient#doExecute(HttpHost, ClassicHttpRequest, HttpContext)}
+ *
+ * @author frankchen
+ */
+public class InternalHttpClient$DoExecute extends AroundInterceptor {
+
+    private static final Set<String> IGNORED_SUFFIXES = new HashSet<>();
+    private final HttpOutgoingMetricsRegistry metricRegistry = HttpOutgoingMetricsRegistry.get();
+
+    public static boolean shouldExclude(String uri) {
+        if (IGNORED_SUFFIXES.isEmpty()) {
+            return false;
+        }
+        String suffix = uri.substring(uri.lastIndexOf(".") + 1).toLowerCase(Locale.ENGLISH);
+        return IGNORED_SUFFIXES.contains(suffix);
+    }
+
+    @Override
+    public InterceptionDecision before(AopContext aopContext) {
+        ClassicHttpRequest httpRequest = aopContext.getArgAs(1);
+        String requestUri = httpRequest.getRequestUri();
+        return shouldExclude(requestUri) ? InterceptionDecision.SKIP_LEAVE : InterceptionDecision.CONTINUE;
+    }
+
+    @Override
+    public void after(AopContext aopContext) {
+        HttpOutgoingMetrics metrics;
+
+        ClassicHttpRequest httpRequest = aopContext.getArgAs(1);
+        String requestUri = httpRequest.getRequestUri();
+        String requestMethod = httpRequest.getMethod();
+
+        if (aopContext.hasException()) {
+            metrics = metricRegistry.addExceptionRequest(requestUri, requestMethod, aopContext.getExecutionTime());
+        } else {
+            metrics = metricRegistry.addRequest(requestUri,
+                                                requestMethod,
+                                                ((CloseableHttpResponse) aopContext.getReturning()).getCode(),
+                                                aopContext.getExecutionTime());
+        }
+
+        /*
+        HttpContext httpContext = aopContext.getArgAs(2);
+        HttpConnection httpConnection = (HttpConnection) (httpContext == null ? null : httpContext.getAttribute("http.connection"));
+        if (httpConnection != null) {
+            try {
+                HttpConnectionMetrics connectionMetrics = httpConnection.getMetrics();
+                metrics.addByteSize(connectionMetrics.getSentBytesCount(), connectionMetrics.getReceivedBytesCount());
+            } catch (ConnectionShutdownException ignored) {
+            }
+        }*/
+    }
+}
